+++ packages/contracts/src/index.ts
@@ -3,4 +3,5 @@ export * from "./domain.js";
 export * from "./errors.js";
 export * from "./idempotency.js";
 export * from "./organization.js";
+export * from "./registry.js";
 export * from "./request-id.js";